requirements.txt
APScheduler==3.9.1
certifi==2022.6.15
charset-normalizer==2.1.0
idna==3.3
install==1.3.5
python-dotenv==0.20.0
pytz==2020.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
schedule==0.6.0
selenium==3.141.0
six==1.16.0
tqdm==4.64.0
tzdata==2022.2
tzlocal==4.2
urllib3==1.25.8
